==> code/02.1_courseinfo_datacleaning.ipynb <==
## 2.1 Course Info Data Cleaning

### Import libraries

# the basics
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

# import preprocessing and modeling
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

# immport metrics
from sklearn.metrics import confusion_matrix, plot_confusion_matrix

pd.options.display.max_colwidth = 350

### Data Cleaning: Independent variables (X)
#### Import data

path = '../data/ds_course_df.csv'
course = pd.read_csv(path)

# quick pre-processing, dropping columns and remove rows with null values
course.drop(columns=['Unnamed: 0'], inplace=True)
course.dropna(inplace=True)

course.shape

course.info()

course.head(1)

### Cleaning up the data in each of the column

#### Remove all non-numeric characters `recent_view`, `num_ratings`, and `length` columns

'''
remove all non-numeric characters from a string
''.join(c for c in "abc123def456" if c.isdigit())

https://stackoverflow.com/questions/1249388/removing-all-non-numeric-characters-from-string-in-python
''.join(c for c in '2,103,753 recent views' if c.isdigit())
'''
# remove all non-numeric characters from a string
course['recent_views'] = course['recent_views'].map(lambda x: int((''.join(c for c in x if c.isdigit()))))
course['num_ratings'] = course['num_ratings'].map(lambda x: int((''.join(c for c in x if c.isdigit()))))
course['length'] = course['length'].map(lambda x: int((''.join(c for c in x if c.isdigit()))))

course.info()

#### Dummify `highlight` into individual features

# example: https://towardsdatascience.com/dealing-with-list-values-in-pandas-dataframes-a177e534f173
# using .apply() and .eval() to turn string into list of items
course['highlight'] = course['highlight'].apply(eval)

# this is a function adopted from https://towardsdatascience.com/dealing-with-list-values-in-pandas-dataframes-a177e534f173
def to_1D(series):
    return pd.Series([x for _list in series for x in _list])

list_highlights = list(to_1D(course['highlight']).unique())

# this is a function adopted from https://towardsdatascience.com/dealing-with-list-values-in-pandas-dataframes-a177e534f173
def boolean_df(item_lists, unique_items):
# this function return a boolean dataframe with the unique items unpacked from the original column.
# Create empty dict
    bool_dict = {}
    
    # Loop through all the tags
    for i, item in enumerate(unique_items):
        
        # Apply boolean mask
        bool_dict[item] = item_lists.apply(lambda x: item in x)
        
    # Return the results as a dataframe
    return pd.DataFrame(bool_dict)

# create the boolean dataframe with only the items in the highlight column
highlight_bool = boolean_df(course['highlight'], list_highlights)

# create a list that include all the highlight items that describe a course is part of a specialization
serials = [col for col in highlight_bool.columns if 'in the' in col] 

# create a new dataframe with only the highlight items that describe a course is part of a specialization
serials_df = highlight_bool[serials]

# sum all columns in one row, and save the result in a new column,
# serials_df['part_of_specialization'] == 1 describe the course is part of a specialization
# sum by row: df["sum"] = df.sum(axis=1)
serials_df['part_of_specialization'] = serials_df.sum(axis=1)
serials_df.head(1)

# create a new column in the highlight data frame to indicate whether a course is part of specialization 
# drop the original itemized columns
highlight_df = highlight_bool.drop(columns=serials)
highlight_df['specialization'] = serials_df['part_of_specialization']

# fix the column names 
highlight_df.columns = highlight_df.columns.str.lower()
highlight_df.columns = highlight_df.columns.str.replace(' ', '_')

for each in highlight_df.columns:
    highlight_df[each] = highlight_df[each].astype(int)

# course.shape
# highlight_df.shape

course = pd.concat([course, highlight_df], axis=1)

#### Clean the `outcome` column

# example: https://towardsdatascience.com/dealing-with-list-values-in-pandas-dataframes-a177e534f173
# using .apply() and .eval() to turn string into list of items
course['outcome'] = course['outcome'].apply(eval)

# define a function to strip the core content of an outcome

def to_core_outcomes(list_of_outcomes):
    coure_string_dict = {}
    
    for each in list_of_outcomes:
        value = ''.join([c for c in each if c.isdigit()])
        key = each.replace(value, '')
        coure_string_dict[key] = int(value)
    return coure_string_dict

# testing 
to_core_outcomes(course['outcome'][0])

course['outcome_core'] = course['outcome'].apply(to_core_outcomes)

course.head(1)

list_outcomes = ['%got a tangible career benefit from this course',
 '%got a pay increase or promotion',
 '%started a new career after completing these courses']

def extract_career_benefit(outcome_dict):
    try:
        return outcome_dict['%got a tangible career benefit from this course']
    except:
        return np.nan

def extract_pay_increase(outcome_dict):
    try:
        return outcome_dict['%got a pay increase or promotion']
    except:
        return np.nan

def extract_new_career(outcome_dict):
    try:
        return outcome_dict['%started a new career after completing these courses']
    except:
        return np.nan

# extract_career_benefit(outcome_bool['outcome_dict'][0])

# course['outcome_dict'].apply(extract_career_benefit)

course['outcome_career_benefit'] = course['outcome_core'].apply(extract_career_benefit)

course['outcome_pay_increase'] = course['outcome_core'].apply(extract_pay_increase)

course['outcome_new_career'] = course['outcome_core'].apply(extract_new_career)

course = course.drop(columns=['highlight','outcome', 'outcome_core'])

#### Create dummies for university partners: `he_partner`
**1: indicating the partner is a university/college vs. companies (i.e. IBM, Google)**

def is_university(partner_name):
    if 'university' in partner_name.lower():
        return 1
    elif 'college' in partner_name.lower():
        return 1
    else:
        return 0

course['he_partner'] = course['partner_title'].apply(is_university)

course.info()

### Data Cleaning: Dependent variables (y)
#### Import data 

enrollment = pd.read_csv('../data/ds_enrollment.csv')
enrollment.drop(columns=['Unnamed: 0'], inplace=True)

enrollment.head()

course = course.merge(enrollment,how='inner', on='course_href')

course.drop_duplicates(inplace=True, ignore_index=True)

course['enrollment'] = course['enrollment'].map(lambda x: int((''.join(c for c in x if c.isdigit()))))

course.shape

# uncomment to run code
# course.to_csv('../data/ds_course_modeling.csv')

# end of codebook
