==> code/02.3_courseinfo_modeling.ipynb <==
## 2.3 Course Information Modeling: predicting enrollment

### Import libraries

# the basics
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

pd.options.display.max_colwidth = 350

#import preprocessing and metrics
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split, cross_val_score, GridSearchCV
from sklearn import metrics
from sklearn.pipeline import Pipeline

# import modeling
from sklearn.linear_model import LinearRegression
from sklearn.neighbors import KNeighborsRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import AdaBoostRegressor
from sklearn.svm import SVR
from xgboost import XGBRegressor

import warnings
warnings.filterwarnings('ignore')

### Import Data and quick cleaning

path = '../data/ds_course_modeling.csv'
course = pd.read_csv(path)

# quick pre-processing, dropping columns and remove rows with null values
course.drop(columns=['Unnamed: 0'], inplace=True)

# check shape
course.shape

# check null values
course.isnull().sum()

# fill all null values with 'median'
course.fillna('median', inplace=True)

course.head(1)

---
## Model Selection

### Baseline Model: mean enrollment (y)

y = course['enrollment']
# null prediction
null_predictions = pd.Series(data=y.mean(), index=y.index)
# null residuals
null_residuals = y - null_predictions

# null RMSE
metrics.mean_squared_error(y, null_predictions, squared=False)

### 1. Modeling: Using all numeric features to predict enrollment.
#### Set up X, y, train-test-split

# get all numeric features
numeric_features = course._get_numeric_data().columns

X = course[numeric_features].drop(columns=['enrollment'])
y = course['enrollment']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

#### Setting up the models in a pipeline and evaluating performance using R2 and RMSE

# Set up a list to contain pipelines
pipelines = []

# Set up tuples for all the models we want to test
models = [('lr', LinearRegression()),
         ('knn', KNeighborsRegressor()),
         ('dtr', DecisionTreeRegressor()),
         ('rfr', RandomForestRegressor()),
         ('ada', AdaBoostRegressor()),
         ('svr', SVR()),
         ('xgb', XGBRegressor())]

# Set up pipelines for all the models alone and with StandardScaler
for model in models:
    pipelines.append(Pipeline([('ss', StandardScaler()),
                              model]))

# this function takes a list of pipelines and return their R2 and RMSE scores 
def pipe_evaluations(pipelines):
    evaluations = []

# Loop through and fit all the pipelines
    for pipe in pipelines:
        pipe.fit(X_train, y_train)

        evaluations.append({
            'Pipeline': [step[0] for step in pipe.steps],
            'R2 (train)': pipe.score(X_train, y_train),
            'R2 (test)': pipe.score(X_test, y_test),
            'RMSE (train)': metrics.mean_squared_error(y_true =y_train, y_pred = pipe.predict(X_train),
                                              squared = False),
            'RMSE (test)': metrics.mean_squared_error(y_true =y_test, y_pred = pipe.predict(X_test),
                                              squared = False)
        })

    # add evaluation metrics to a dataframe
    evaluations_df = pd.DataFrame(evaluations)
    return evaluations_df

# evaluation the pipelines
pipe_evaluations(pipelines)

#### Shortlisting models based on performance

**Takeaways:** based on the test model performance (R2 scores and RMSE), the two models were selected for GridSearch.
- **Decision Tree**
- **XGBoost**


**In addition,** for interpretability, I will also continue exploring the **Linear Regression** model and will apply polynominal features and regularization as next steps.

### 2. Excluding `100%_online`, `flexible_deadlines` and all the 'language labels from X
#### Set up X, y

X = course[numeric_features].drop(columns=['spanish','chinese_(traditional)', 'arabic', 'portuguese_(brazilian)', 
                                           'russian', 'advanced_level', 'chinese_(simplified)', 'french', 
                                           'japanese', '100%_online', 'flexible_deadlines', 'enrollment'])
y = course['enrollment']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

#### Setting up the models in a pipeline and evaluating performance using R2 and RMSE

# Set up a list to contain pipelines
pipelines = []

# Set up tuples for all the models we want to test
models = [('lr', LinearRegression()),
         ('knn', KNeighborsRegressor()),
         ('dtr', DecisionTreeRegressor()),
         ('rfr', RandomForestRegressor()),
         ('ada', AdaBoostRegressor()),
         ('svr', SVR()),
         ('xgb', XGBRegressor())]

# Set up pipelines for all the models alone and with StandardScaler
for model in models:
    pipelines.append(Pipeline([('ss', StandardScaler()),
                              model]))

# evaluation the pipelines
pipe_evaluations(pipelines)

**Takeaways:** based on the test model performance (R2 scores and RMSE), the two models were selected for GridSearch.
- **Decision Tree**
- **XGBoost**


For interpretability, I will also continue exploring the **Linear Regression** model and will apply polynominal features and regularization as next steps. **In addition,** Neural Networks are also explored as a benchmark to compare model performance.

---
## Tune shortlisted models, using the X excluding `100%_online`, `flexible_deadlines` and all the 'language labels  
Shortlisted models:
- **Decision Tree** (tuning using GridSearch)
- **XGBoost** (tuning using GridSearch??)
- **Linear Regression** (tuning using polynormial features and regularization)

from sklearn.model_selection import GridSearchCV

X = course[numeric_features].drop(columns=['spanish','chinese_(traditional)', 'arabic', 'portuguese_(brazilian)', 
                                           'russian', 'advanced_level', 'chinese_(simplified)', 'french', 
                                           'japanese', '100%_online', 'flexible_deadlines', 'enrollment'])
y = course['enrollment']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

### GridSearch - Decision Tree Regressor

# without tuning
pipe = Pipeline([('ss', StandardScaler()),
                  ('dtr', DecisionTreeRegressor())])


pipe.fit(X_train, y_train)
pipe.score(X_train, y_train), pipe.score(X_test, y_test)

# set up the pipe
pipe = Pipeline([('ss', StandardScaler()),
                  ('dtr', DecisionTreeRegressor())])

# Create dictionary with candidate learning algorithms and their hyperparameters
grid_param = [{'dtr__max_depth':[11,12,13,None],
               'dtr__min_samples_split':[4,6,8],
               'dtr__min_samples_leaf': [1,2,3],
               'dtr__max_leaf_nodes': [10,11,12,None]}]

# create a gridsearch of the pipeline, and fit the best model
gridsearch = GridSearchCV(pipe, grid_param, cv=5) 
best_model = gridsearch.fit(X_train,y_train)

# print metrics
print(best_model.best_estimator_)
print('')
print(best_model.score(X_train, y_train), best_model.score(X_test, y_test))

**Key takeways:** R-square score improves from **0.820 to 0.838** with GridSearching the Decision Tree Model.

### GridSearch - XGBoost

**Performance without GridSearch:** Before tuning XGBoost model, take a look at the performance of the default setting. There is overfitting in the model and I will try to tune it by adding some bias.

# without tuning
pipe = Pipeline([('ss', StandardScaler()),
                  ('xgb', XGBRegressor())])


pipe.fit(X_train, y_train)
pipe.score(X_train, y_train), pipe.score(X_test, y_test)

# Create dictionary with candidate learning algorithms and their hyperparameters
grid_param = [{'xgb__max_depth':[4,5],
               'xgb__min_child_weight':[1,2],
               'xgb__n_estimators': [100,200],
               'xgb__learning_rate': [0.2,0.3,0.4]}]

# create a gridsearch of the pipeline, and fit the best model
gridsearch = GridSearchCV(pipe, grid_param, cv=5) 
best_model = gridsearch.fit(X_train,y_train)

print(best_model.best_estimator_)
print('')
print(best_model.score(X_train, y_train), best_model.score(X_test, y_test))

**Key takeways:** R-square score improves from **0.863 to 0.885** with GridSearching the XGBoost Model.

### Linear Regression and Regularization

from sklearn.preprocessing import PolynomialFeatures
from sklearn.linear_model import Ridge, RidgeCV
from sklearn.linear_model import Lasso, LassoCV

# Recall:
X = course[numeric_features].drop(columns=['spanish','chinese_(traditional)', 'arabic', 'portuguese_(brazilian)', 
                                           'russian', 'advanced_level', 'chinese_(simplified)', 'french', 
                                           'japanese', '100%_online', 'flexible_deadlines', 'enrollment'])
y = course['enrollment']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

#### Linear Regression without regularization

pipe = Pipeline([('ss', StandardScaler()),
                 ('lr', LinearRegression())])

pipe.fit(X_train, y_train)
pipe.score(X_train, y_train), pipe.score(X_test, y_test)

#### LASSO with PolynomialFeatures

# set up the pipe
pipe = Pipeline([('poly', PolynomialFeatures()),
                  ('ss', StandardScaler()),
                  ('lasso', Lasso())])

pipe.fit(X_train, y_train)
pipe.score(X_train, y_train), pipe.score(X_test, y_test)

# Create dictionary with candidate learning algorithms and their hyperparameters
grid_param = [{'poly__degree':[2,3],
               'poly__interaction_only':[True, False],
               'lasso__alpha': [1, 10, 100]}]

# create a gridsearch of the pipeline, and fit the best model
gridsearch = GridSearchCV(pipe, grid_param, cv=5) 
best_model = gridsearch.fit(X_train,y_train)

print(best_model.best_estimator_)
print('')
print(best_model.score(X_train, y_train), best_model.score(X_test, y_test))

**Key takeways:** R-square score improves from **0.581 to 0.883** with LASSO Regularization.

#### Ridge with PolynomialFeatures"

pipe = Pipeline([('poly', PolynomialFeatures()),
                  ('ss', StandardScaler()),
                  ('ridge', Ridge())])

pipe.fit(X_train, y_train)
pipe.score(X_train, y_train), pipe.score(X_test, y_test)

# Create dictionary with candidate learning algorithms and their hyperparameters
grid_param = [{'poly__degree':[2,3],
               'poly__interaction_only':[True, False],
               'ridge__alpha': [0.01,0.1,1]}]

# create a gridsearch of the pipeline, and fit the best model
gridsearch = GridSearchCV(pipe, grid_param, cv=5) 
best_model = gridsearch.fit(X_train,y_train)

print(best_model.best_estimator_)
print('')
print(best_model.score(X_train, y_train), best_model.score(X_test, y_test))

**Key takeways:** R-square score improves from **0.581 to 0.868** with Ridge Regularization.

#### Using LASSO for feature selection and intepretation

from sklearn.feature_selection import SelectFromModel
# reference:
# https://towardsdatascience.com/feature-selection-using-regularisation-a3678b71e499

poly = PolynomialFeatures(degree=3, interaction_only=False)
X_train_overfit = poly.fit_transform(X_train)
X_test_overfit = poly.transform(X_test)
sc = StandardScaler()
X_train_overfit_sc = sc.fit_transform(X_train_overfit)
X_test_overfit_sc = sc.transform(X_test_overfit)
lasso_lr = Lasso(alpha=7500)
lasso_lr.fit(X_train_overfit_sc, y_train)
lasso_lr.score(X_train_overfit_sc, y_train), lasso_lr.score(X_test_overfit_sc, y_test)

# can also use SelectFromModel 

# from sklearn.feature_selection import SelectFromModel
# reference:
# https://towardsdatascience.com/feature-selection-using-regularisation-a3678b71e499

# sel_ = SelectFromModel(Lasso(alpha=3000))
# sel_.fit(X_train_overfit_sc, y_train)

# put the selected features into a DataFrame
selections = pd.DataFrame(data=lasso_lr.coef_, index=poly.get_feature_names(X_train.columns), columns=['model_coefficient'])

# list the features of the non-zero coefficients 
selections[selections['model_coefficient']!=0]

# get the feature names of the non-zero features
selections[selections['model_coefficient']!=0].index

**Key Takeways**:  
The most important features are: `'num_ratings', 'num_reviews', 'num_ratings length', 'num_ratings shareable_certificate', 'num_ratings he_partner', 'num_reviews shareable_certificate', 'length shareable_certificate', 'num_ratings shareable_certificate he_partner', 'num_ratings he_partner^2', 'num_reviews^2 beginner_level', 'num_reviews shareable_certificate^2', 'length^2 he_partner', 'length shareable_certificate^2'`   
All these features and their poly/interaction terms are positively related to enrollment except for `num_reviews^2 beginner_level`. 

# end of notebook
