==> code/03.2_coursereview_topic_modeling.ipynb <==
## 3.2 Course Review Topic Modeling

**Credit to Analytics Vidhya Articles for the ideas and codes:**  
https://www.analyticsvidhya.com/blog/2018/10/mining-online-reviews-topic-modeling-lda/  
https://www.analyticsvidhya.com/blog/2016/08/beginners-guide-to-topic-modeling-in-python/

### Import libraries

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

import re
import gensim
from gensim import corpora
import nltk
from nltk import FreqDist

%matplotlib inline
pd.options.display.max_colwidth = 350

---
### Course Name: Review from 'Machine Learning by Stanford University, Andrew Ng' on Coursera
### Data Import

path = './data/reviews_machine_learning.csv'
reviews = pd.read_csv(path)

# quick pre-processing, dropping columns and remove rows with null values
reviews.drop(columns=['Unnamed: 0'], inplace=True)
reviews.dropna(inplace=True)

# create review label: 1 - 3: bad reviews, 4/5: good reviews
reviews['review_is_bad'] = reviews['rating'].map(lambda x: 1 if x < 4 else 0)

### Data Pre-processing 
#### `tokenize`, `stopwords`, `lemmatize`, `pos_tag`

from nltk.corpus import wordnet

# define a function to associate the wordnet object value corresponding to the POS tag
def get_wordnet_pos(pos_tag):
    if pos_tag.startswith('J'):
        return wordnet.ADJ
    elif pos_tag.startswith('V'):
        return wordnet.VERB
    elif pos_tag.startswith('N'):
        return wordnet.NOUN
    elif pos_tag.startswith('R'):
        return wordnet.ADV
    else:
        return wordnet.NOUN

import string
from nltk import pos_tag
from nltk.corpus import stopwords
from nltk.tokenize import WhitespaceTokenizer
from nltk.stem import WordNetLemmatizer

def clean_text(text):
    # lower text
    text = text.lower()
    # tokenize text and remove puncutation
    text = [word.strip(string.punctuation) for word in text.split(" ")]
    # remove words that contain numbers
    text = [word for word in text if not any(c.isdigit() for c in word)]
    # remove stop words
    stop = stopwords.words('english')
    text = [x for x in text if x not in stop]
    # remove empty tokens
    text = [t for t in text if len(t) > 0]
    # pos tag text
    pos_tags = pos_tag(text)
    # lemmatize text
    text = [WordNetLemmatizer().lemmatize(t[0], get_wordnet_pos(t[1])) for t in pos_tags]
    # remove words with only one letter
    text = [t for t in text if len(t) > 1]
    # join all
    text = " ".join(text)
    return(text)

# clean text data
reviews["review_clean"] = reviews["review"].apply(lambda x: clean_text(x))

reviews.head(1)

## LDA Modeling

from nltk import word_tokenize

def nouns_adj(text):
    # Given a string of text, tokenize the text and pull out only the nouns and adjectives
    is_noun_adj = lambda pos: pos[:2] == 'NN' or pos[:2] == 'JJ'
    tokenized = word_tokenize(text)
    nouns_adj = [word for (word, pos) in pos_tag(tokenized) if   is_noun_adj(pos)]
    return ' '.join(nouns_adj)

reviews["review_lda"] = reviews["review_clean"].apply(lambda x: nouns_adj(x))

lda_reviews = reviews["review_lda"]

from sklearn.feature_extraction.text import CountVectorizer

# create document term matrix
min_df = 4 # minimum required occurences of a word, e.g 4
max_features = 10000 # max number of unique words, e.g 10000

vectorizer = CountVectorizer(
              min_df=min_df, 
              max_features=max_features, 
              ngram_range=(1,2))

vec_lda_reviews = vectorizer.fit_transform(lda_reviews.values)

vec_lda_reviews_df = pd.DataFrame(vec_lda_reviews.toarray(), columns=vectorizer.get_feature_names())
vec_lda_reviews_df.index = lda_reviews.index
vec_lda_reviews_df.head()

tokenized_reviews = pd.Series(lda_reviews).apply(lambda x: x.split())

dictionary = corpora.Dictionary(tokenized_reviews)

doc_term_matrix = [dictionary.doc2bow(rev) for rev in tokenized_reviews]

# Creating the object for LDA model using gensim library
LDA = gensim.models.ldamodel.LdaModel

# Build LDA model
lda_model = LDA(corpus=doc_term_matrix, id2word=dictionary, num_topics=6, random_state=42,
                chunksize=1000, passes=50)

lda_model.print_topics()

import pyLDAvis
import pyLDAvis.gensim

# Visualize the topics
pyLDAvis.enable_notebook()
vis = pyLDAvis.gensim.prepare(lda_model, doc_term_matrix, dictionary)
vis
